src/requirements.txt
pymongo
python-dotenv
flask
azure-keyvault-secrets
azure-identity
